+++ source/reject_files_larger_than_original/changelog.md
@@ -1,3 +1,7 @@
 
+**<span style="color:#56adda">0.0.2</span>**
+- Add postprocessor and library file test runners to mark a file as failed and prevent it from being re-added
+- Fix file copy issue for worker runner
+
 **<span style="color:#56adda">0.0.1</span>**
 - Initial version